Renaming of the 'constraints' folder to 'formulae'
We have a Constraint class which is a very different beast from the formulae in the old constraints folder. To make sure this naming is not confusing, I have renamed 'constraints' to 'formulae' to make it clear that the contents of that folder are meant to act as the formulae to construct Constraint instances from.